pyproject.toml
[project]
name = "raspi-pico"
version = "0.1.0"
description = "這是我使用樹莓派與pico整合的專案"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "flask-socketio>=5.5.1",
    "ipykernel>=7.1.0",
    "openpyxl>=3.1.5",
    "paho-mqtt>=2.1.0",
]